# perf: reduce query when calling `get_doc` with filters