When calling `get_doc` in the format `get_doc(doctype, filters)`, an additional query was run to ascertain the name of the document. This query is unnecessary because the query run in `load_from_db` immediately after accomplishes the same result.

### Other Changes
- Remove unused property `_default_new_docs`
- Move `self.doctype = args[0]` outside if-else block for readability

---

### Before

```python

In [1]: %timeit frappe.get_doc("UOM", {})
456 µs ± 41.1 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

### After (30% improvement)

```python
In [1]: %timeit frappe.get_doc("UOM", {})
311 µs ± 10.1 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```