@sagarvora 2022-03-31
Added a second commit which reduces unnecessary conditions and improves readability. Additionally it aliases `reload` in a one-liner instead of existing definition.